add responsive, change structure if file
