+++ src/components/HeaderAndFooter/Header/Drawer.scss
@@ -1,0 +1,55 @@
+@import "../../../sass_config/color.scss";
+@import "../../../sass_config/mixin.scss";
+
+.logo-closebtn-wrapper {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  height: 108px;
+  padding: 0 8px;
+  .company-logo {
+    height: 60px;
+    cursor: pointer;
+  }
+  .close-icon {
+    cursor: pointer;
+    font-size: 40px;
+    color: $gray;
+  }
+}
+.nav-bar-drawer {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  .menu-drawer {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: center;
+    margin-bottom: 20px;
+
+    .menu-item {
+      margin: 16px;
+      padding: 8px 0;
+
+      &:hover {
+        color: $color_primary;
+        text-shadow: 0 0.015em $color_primary, 0 -0.015em$color_primary,
+          0.01em 0 $color_primary, -0.01em 0 $color_primary;
+      }
+    }
+  }
+  .round-button-drawer {
+    border: 1.8px solid $color_primary;
+    border-radius: 100px;
+    padding: 8px 32px;
+    margin: 8px;
+    width: 90vw;
+    font-weight: 700;
+  }
+  .background-white-drawer {
+    background-color: white;
+    color: $color_primary;
+  }
+}
